Fixes for commit model. New web site now support running against commit model.

+++ src/Codex.ElasticSearch/Store/ElasticSearchStoredFilterBuilder.cs
@@ -152,7 +152,7 @@ namespace Codex.ElasticSearch
         public ElasticEntityRef(ISearchEntity entity)
         {
             Uid = entity.Uid;
-            StableId = (int)entity.StableId;
+            StableId = entity.StableId;
         }
 
         public override string ToString()
